Add batch

+++ app/Http/Controllers/PlanController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 use App\Enums\ReadingPlanStatus;
 use Illuminate\Http\Request;
-use App\Models\ReadingPlan;
+use App\Models\Plan;
 use App\Models\Book;
-class ReadingplanController extends Controller
+use App\Http\Requests\PlanRequest;
+
+class PlanController extends Controller
 {
     public function index(Request $request){
 
         $currentStatus = $request->status;
 
         $readingPlans =auth()->user()
-        ->readingPlans()
+        ->Plans()
         ->when($currentStatus, function ($query) use ($currentStatus) {
             $query->where('status', $currentStatus);
         })
@@ -28,8 +30,8 @@ class ReadingplanController extends Controller
         return view('reading-plans.create',compact('books'));
     }
 
-    public function store(Request $request){
-        $book = ReadingPlan::create([
+    public function store(PlanRequest $request,Plan $plan){
+        $book = Plan::create([
             'user_id'=>auth()->id(),
             'book_id'=>$request->book_id,
             'target_date'=>$request->target_date, 
@@ -40,33 +42,32 @@ class ReadingplanController extends Controller
         ->route('reading-plans.index');
     }
 
-    public function complete(Request $request,$id){
+    public function complete(Request $request,Plan $plan){
+        $this->authorize('update', $plan);
 
-        $readingPlan = ReadingPlan::findOrFail($id);
-        $readingPlan->update([
+        $plan->update([
             'status'=> ReadingPlanStatus::Completed->value,
         ]);
 
         return redirect()->route('reading-plans.index');
     }
 
-    public function edit(Request $request,$id){
-        $readingPlan = ReadingPlan::findOrFail($id);
+    public function edit(PlanRequest $request,Plan $plan){
+        $this->authorize('update', $plan);
         return view('reading-plans.edit',compact('readingPlan'));
     }
 
-    public function update(Request $request,$id){
-
-        $readingPlan = ReadingPlan::findOrFail($id);
-        $readingPlan->update([
+    public function update(PlanRequest $request,Plan $plan){
+        $this->authorize('update', $plan);
+        $plan->update([
             'target_date'=>$request->target_date,
         ]);
 
         return redirect()->route('reading-plans.index');
     }
 
-    public function destroy(ReadingPlan $plan){
-    
+    public function destroy(Plan $plan){
+        $this->authorize('delete', $plan);
         $plan->delete();
         return redirect()->route('reading-plans.index');
     }
